+++ Samples/TrendingMovies/TrendingMovies/TMDb/Movies.swift
@@ -1,0 +1,7 @@
+/// https://developers.themoviedb.org/3/movies/get-now-playing
+struct Movies: Codable {
+    let page: Int
+    let results: [Movie]
+    let totalPages: Int
+    let totalResults: Int
+}