+++ Sources/Sentry/include/SentryCrashIntegration.h
@@ -3,9 +3,13 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+@class SentryScope, SentryCrashWrapper;
+
 static NSString *const SentryDeviceContextFreeMemoryKey = @"free_memory";
 
 @interface SentryCrashIntegration : NSObject <SentryIntegrationProtocol>
+
++ (void)enrichScope:(SentryScope *)scope crashWrapper:(SentryCrashWrapper *)crashWrapper;
 
 @end