+++ Tests/SentryTests/Performance/SentryTracerTests.swift
@@ -6,6 +6,7 @@ class SentryTracerTests: XCTestCase {
         let client: TestClient
         let hub: TestHub
         let scope: Scope
+        let dispatchQueue = TestSentryDispatchQueueWrapper()
         
         let transactionName = "Some Transaction"
         let transactionOperation = "ui.load"
@@ -18,6 +19,8 @@ class SentryTracerTests: XCTestCase {
         let appStart: Date
         let appStartEnd: Date
         let appStartDuration = 0.5
+        
+        let idleTimeout: TimeInterval = 1.0
         
         #if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
         var displayLinkWrapper: TestDiplayLinkWrapper
@@ -56,6 +59,10 @@ class SentryTracerTests: XCTestCase {
         
         func getSut(waitForChildren: Bool = true) -> SentryTracer {
             return hub.startTransaction(with: transactionContext, bindToScope: false, waitForChildren: waitForChildren, customSamplingContext: [:]) as! SentryTracer
+        }
+        
+        func getSut(idleTimeout: TimeInterval = 0.0, dispatchQueueWrapper: SentryDispatchQueueWrapper) -> SentryTracer {
+            return hub.startTransaction(with: transactionContext, bindToScope: false, customSamplingContext: [:], idleTimeout: idleTimeout, dispatchQueueWrapper: dispatchQueueWrapper)
         }
     }
     
@@ -123,6 +130,171 @@ class SentryTracerTests: XCTestCase {
         sut.finish()
         
         XCTAssertEqual(0, fixture.hub.capturedEventsWithScopes.count)
+    }
+    
+    func testIdleTimeout_NoChildren_TransactionNotCaptured() {
+        _ = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        
+        fixture.dispatchQueue.invokeLastDispatchAfter()
+        
+        XCTAssertEqual(0, fixture.hub.capturedEventsWithScopes.count)
+    }
+    
+    func testIdleTimeout_InvokesDispatchAfterWithCorrectWhen() {
+        _ = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        
+        fixture.dispatchQueue.invokeLastDispatchAfter()
+        
+        let expectedWhen = fixture.currentDateProvider.dispatchTimeNow() + UInt64(fixture.idleTimeout) * NSEC_PER_SEC
+        XCTAssertEqual(expectedWhen, fixture.dispatchQueue.dispatchAfterInvocations.invocations.first?.when)
+    }
+    
+    func testIdleTimeout_SpanAdded_IdleTimeoutCancelled() {
+        let sut = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        
+        sut.startChild(operation: fixture.transactionOperation)
+        
+        XCTAssertEqual(1, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        XCTAssertEqual(1, fixture.dispatchQueue.dispatchCancelInvocations.count)
+    }
+    
+    func testIdleTimeoutWithRealDispatchQueue_SpanAdded_IdleTimeoutCancelled() {
+        let sut = fixture.getSut(idleTimeout: 0.1, dispatchQueueWrapper: SentryDispatchQueueWrapper())
+        
+        let child = sut.startChild(operation: fixture.transactionOperation)
+        let grandChild = child.startChild(operation: fixture.transactionOperation)
+        grandChild.finish()
+        child.finish()
+        
+        delayNonBlocking(timeout: 0.5)
+        
+        assertOneTransactionCaptured(sut)
+    }
+    
+    func testIdleTimeout_TwoChildren_FirstFinishes_WaitsForTheOther() {
+        let sut = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        
+        let child1 = sut.startChild(operation: fixture.transactionOperation)
+        let child2 = sut.startChild(operation: fixture.transactionOperation)
+        
+        XCTAssertEqual(1, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        child1.finish()
+        XCTAssertEqual(1, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        
+        child2.finish()
+        XCTAssertEqual(2, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        
+        fixture.dispatchQueue.invokeLastDispatchAfter()
+        
+        assertOneTransactionCaptured(sut)
+    }
+    
+    func testIdleTimeout_ChildSpanFinished_IdleStarted() {
+        let sut = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        XCTAssertEqual(1, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        
+        let child = sut.startChild(operation: fixture.transactionOperation)
+        XCTAssertEqual(1, fixture.dispatchQueue.dispatchCancelInvocations.count)
+        
+        child.finish()
+        XCTAssertEqual(2, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        
+        let grandChild = child.startChild(operation: fixture.transactionOperation)
+        XCTAssertEqual(3, fixture.dispatchQueue.dispatchCancelInvocations.count)
+        
+        grandChild.finish()
+        XCTAssertEqual(3, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        XCTAssertEqual(4, fixture.dispatchQueue.dispatchCancelInvocations.count)
+        
+        fixture.dispatchQueue.invokeLastDispatchAfter()
+        
+        assertOneTransactionCaptured(sut)
+    }
+    
+    func testIdleTimeout_TimesOut_TrimsEndTimestamp() {
+        let sut = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        
+        let child1 = sut.startChild(operation: fixture.transactionOperation)
+        advanceTime(bySeconds: 1.0)
+        child1.finish()
+        
+        let child2 = sut.startChild(operation: fixture.transactionOperation)
+        advanceTime(bySeconds: 1.0)
+        let expectedEndTimestamp = fixture.currentDateProvider.date()
+        child2.finish()
+        
+        advanceTime(bySeconds: fixture.idleTimeout)
+        fixture.dispatchQueue.invokeLastDispatchAfter()
+        
+        assertOneTransactionCaptured(sut)
+        XCTAssertEqual(expectedEndTimestamp, sut.timestamp)
+    }
+    
+    func testIdleTimeout_CallFinish_TrimsEndTimestamp() {
+        let sut = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        
+        let child = sut.startChild(operation: fixture.transactionOperation)
+        advanceTime(bySeconds: 1.0)
+        child.finish()
+        let expectedEndTimestamp = fixture.currentDateProvider.date()
+        
+        advanceTime(bySeconds: 1.0)
+        sut.finish(status: .cancelled)
+        
+        assertOneTransactionCaptured(sut)
+        XCTAssertEqual(expectedEndTimestamp, sut.timestamp)
+    }
+    
+    func testNonIdleTransaction_CallFinish_DoesNotTrimEndTimestamp() {
+        let sut = fixture.getSut()
+        
+        advanceTime(bySeconds: 1.0)
+        let child = sut.startChild(operation: fixture.transactionOperation)
+        child.finish()
+        advanceTime(bySeconds: 1.0)
+        
+        let expectedEndTimestamp = fixture.currentDateProvider.date()
+        sut.finish()
+        
+        assertOneTransactionCaptured(sut)
+        XCTAssertEqual(expectedEndTimestamp, sut.timestamp)
+    }
+    
+    func testIdleTimeoutWithUnfinishedChildren_TimesOut_TrimsEndTimestamp() {
+        let sut = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        
+        let child1 = sut.startChild(operation: fixture.transactionOperation)
+        advanceTime(bySeconds: 1.0)
+        child1.finish()
+        advanceTime(bySeconds: 1.0)
+        _ = sut.startChild(operation: fixture.transactionOperation)
+        
+        advanceTime(bySeconds: fixture.idleTimeout)
+        
+        fixture.dispatchQueue.invokeLastDispatchAfter()
+        let expectedEndTimestamp = fixture.currentDateProvider.date()
+        
+        assertOneTransactionCaptured(sut)
+        XCTAssertEqual(expectedEndTimestamp, sut.timestamp)
+    }
+    
+    func testIdleTimeout_CallFinish_WaitsForChildren_DoesntStartTimeout() {
+        let sut = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        
+        let child = sut.startChild(operation: fixture.transactionOperation)
+        XCTAssertEqual(1, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        
+        sut.finish()
+        XCTAssertEqual(2, fixture.dispatchQueue.dispatchCancelInvocations.count)
+        XCTAssertEqual(1, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        
+        XCTAssertFalse(sut.isFinished)
+        advanceTime(bySeconds: 1)
+        child.finish()
+        
+        XCTAssertEqual(1, fixture.dispatchQueue.dispatchAfterInvocations.count)
+        
+        XCTAssertTrue(sut.isFinished)
     }
     
     func testAddColdAppStartMeasurement_PutOnNextAutoUITransaction() {
@@ -261,6 +433,24 @@ class SentryTracerTests: XCTestCase {
         XCTAssertTrue(child3.isFinished)
         XCTAssertEqual(child3.context.status, .deadlineExceeded)
         XCTAssertEqual(sut.timestamp, child3.timestamp)
+    }
+    
+    func testFinishCallback_CalledWhenTracerFinishes() {
+        let callbackExpectation = expectation(description: "FinishCallback called")
+        
+        let sut = fixture.getSut(idleTimeout: fixture.idleTimeout, dispatchQueueWrapper: fixture.dispatchQueue)
+        
+        let block: (SentryTracer) -> Void = { tracer in
+            XCTAssertEqual(sut, tracer)
+            callbackExpectation.fulfill()
+        }
+        sut.finishCallback = block
+        
+        fixture.dispatchQueue.invokeLastDispatchAfter()
+        
+        XCTAssertNil(sut.finishCallback)
+        
+        wait(for: [callbackExpectation], timeout: 0.1)
     }
     
     // Although we only run this test above the below specified versions, we expect the
@@ -431,6 +621,14 @@ class SentryTracerTests: XCTestCase {
     }
     #endif
     
+    private func advanceTime(bySeconds: TimeInterval) {
+        fixture.currentDateProvider.setDate(date: fixture.currentDateProvider.date().addingTimeInterval(bySeconds))
+        
+        let delta = bySeconds * Double(NSEC_PER_SEC)
+        let newNow = fixture.currentDateProvider.internalDispatchNow + .nanoseconds(Int(delta))
+        fixture.currentDateProvider.internalDispatchNow = newNow
+    }
+    
     private func getSerializedTransaction() -> [String: Any] {
         guard let transaction = fixture.hub.capturedEventsWithScopes.first?.event else {
             fatalError("Event must not be nil.")
@@ -504,5 +702,5 @@ class SentryTracerTests: XCTestCase {
         let serializedTransaction = fixture.hub.capturedEventsWithScopes.first!.event.serialize()
         XCTAssertNil(serializedTransaction["measurements"])
     }
-
+    
 }