+++ Tests/SentryTests/SentryHubTests.swift
@@ -202,12 +202,12 @@ class SentryHubTests: XCTestCase {
     }
     
     func testCaptureEventWithoutScope() {
-        fixture.getSut().capture(event: fixture.event)
+        fixture.getSut(fixture.options, fixture.scope).capture(event: fixture.event)
         
         XCTAssertEqual(1, fixture.client.captureEventWithScopeInvocations.count)
         if let eventArguments = fixture.client.captureEventWithScopeInvocations.first {
             XCTAssertEqual(fixture.event.eventId, eventArguments.event.eventId)
-            XCTAssertEqual(Scope(), eventArguments.scope)
+            XCTAssertEqual(fixture.scope, eventArguments.scope)
         }
     }
     
@@ -377,12 +377,12 @@ class SentryHubTests: XCTestCase {
     }
     
     func testCaptureMessageWithoutScope() {
-        fixture.getSut().capture(message: fixture.message)
+        fixture.getSut(fixture.options, fixture.scope).capture(message: fixture.message)
         
         XCTAssertEqual(1, fixture.client.captureMessageWithScopeInvocations.count)
         if let messageArguments = fixture.client.captureMessageWithScopeInvocations.first {
             XCTAssertEqual(fixture.message, messageArguments.message)
-            XCTAssertEqual(Scope(), messageArguments.scope)
+            XCTAssertEqual(fixture.scope, messageArguments.scope)
         }
     }
     
@@ -416,12 +416,12 @@ class SentryHubTests: XCTestCase {
     }
     
     func testCatpureErrorWithoutScope() {
-        fixture.getSut().capture(error: fixture.error).assertIsNotEmpty()
+        fixture.getSut(fixture.options, fixture.scope).capture(error: fixture.error).assertIsNotEmpty()
         
         XCTAssertEqual(1, fixture.client.captureErrorWithScopeInvocations.count)
         if let errorArguments = fixture.client.captureErrorWithScopeInvocations.first {
             XCTAssertEqual(fixture.error, errorArguments.error as NSError)
-            XCTAssertEqual(Scope(), errorArguments.scope)
+            XCTAssertEqual(fixture.scope, errorArguments.scope)
         }
     }
     
@@ -436,12 +436,12 @@ class SentryHubTests: XCTestCase {
     }
     
     func testCatpureExceptionWithoutScope() {
-        fixture.getSut().capture(exception: fixture.exception).assertIsNotEmpty()
+        fixture.getSut(fixture.options, fixture.scope).capture(exception: fixture.exception).assertIsNotEmpty()
         
         XCTAssertEqual(1, fixture.client.captureExceptionWithScopeInvocations.count)
         if let errorArguments = fixture.client.captureExceptionWithScopeInvocations.first {
             XCTAssertEqual(fixture.exception, errorArguments.exception)
-            XCTAssertEqual(Scope(), errorArguments.scope)
+            XCTAssertEqual(fixture.scope, errorArguments.scope)
         }
     }
     
@@ -799,14 +799,30 @@ class SentryHubTests: XCTestCase {
         XCTAssertNotEqual(SentryId.empty, SentryId(uuidString: profile["profile_id"] as! String))
         XCTAssertNotEqual(SentryId.empty, SentryId(uuidString: profile["trace_id"] as! String))
         
-        XCTAssertFalse(((profile["debug_meta"] as! [String: Any])["images"] as! [Any]).isEmpty)
+        let images = (profile["debug_meta"] as! [String: Any])["images"] as! [[String: Any]]
+        XCTAssertFalse(images.isEmpty)
+        let firstImage = images[0]
+        XCTAssertFalse((firstImage["code_file"] as! String).isEmpty)
+        XCTAssertFalse((firstImage["debug_id"] as! String).isEmpty)
+        XCTAssertFalse((firstImage["image_addr"] as! String).isEmpty)
+        XCTAssertGreaterThan((firstImage["image_size"] as! Int), 0)
+        XCTAssertEqual(firstImage["type"] as! String, "macho")
+        
         let sampledProfile = profile["sampled_profile"] as! [String: Any]
+        let threadMetadata = sampledProfile["thread_metadata"] as! [String: Any]
+        let queueMetadata = sampledProfile["queue_metadata"] as! [String: Any]
+        
+        XCTAssertFalse(threadMetadata.isEmpty)
+        XCTAssertGreaterThan((threadMetadata.first?.value as! [String: Any])["priority"] as! Int, 0)
+        XCTAssertFalse(queueMetadata.isEmpty)
+        XCTAssertFalse(((queueMetadata.first?.value as! [String: Any])["label"] as! String).isEmpty)
+        
         let samples = sampledProfile["samples"] as! [[String: Any]]
         XCTAssertFalse(samples.isEmpty)
-        
         let frames = samples[0]["frames"] as! [[String: Any]]
         XCTAssertFalse(frames.isEmpty)
         XCTAssertFalse((frames[0]["instruction_addr"] as! String).isEmpty)
+        XCTAssertFalse((frames[0]["function"] as! String).isEmpty)
     }
     
     private func testSampler(expected: SentrySampleDecision, options: (Options) -> Void) {