+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -161,4 +161,6 @@
 
 #if SENTRY_HAS_UIKIT
 #    import "SentryANRTrackingIntegration.h"
+#    import "SentryUIEventTracker.h"
+#    import "SentryUIEventTrackingIntegration.h"
 #endif