+++ CHANGELOG.md
@@ -2,9 +2,22 @@
 
 ## Unreleased
 
+### Fixes
+
+- Fix JSON conversion error message (#1856)
+
+## 7.16.0
+
 ### Features
 
+- UI event transactions for clicks (#1784)
+- Collect queue label information for profiles (#1828)
+- Use the macho format for debug information in Profiling (#1830)
 - Allow partial SDK info override (#1816)
+
+### Fixes
+
+- Hub uses its scope (#1821)
 
 ## 7.15.0